+++ docs/CHANGELOG.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* feature : add templator for resolving dashboard uids on alerts
 * fix : idempotency of bulk ops importing alerts
 * fix : do not double-decode path (mostly affects bulk ops)
 * feature : cli supports extra option to remove edit metadata from dashboards